Add widget MdDevicesSelector
